prompt_tuning/requirements.txt
# Prompt Tuning 實驗環境依賴
google-genai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pyaudio>=0.2.14
